requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyYAML==6.0.1
python-dotenv==1.0.0
groq==0.3.2
streamlit==1.32.0
plotly==5.18.0
pandas==2.2.0
